// src/types/api.ts
export interface components {
    schemas: {
// ...
        Body_transcribe_file_v1_audio_transcriptions_post: {
            /** Model */
            model: string;
            /** Language */
            language?: string | null;
            /** Prompt */
            prompt?: string | null;
            /** @default json */
            response_format: components["schemas"]["ResponseFormat"];
            /**
             * Temperature
             * @default 0
             */
            temperature: number;
            /**
             * Timestamp Granularities
             * @default [
             *       "segment"
             *     ]
             */
            timestamp_granularities: ("segment" | "word")[];
            /**
             * Stream
             * @default false
             */
            stream: boolean;
            /** Hotwords */
            hotwords?: string | null;
            /**
             * Without Timestamps
             * @default true
             */
            without_timestamps: boolean;
            /**
             * File
             * Format: binary
             */
            file: string;
        };
        /** Body_translate_file_v1_audio_translations_post */
        Body_translate_file_v1_audio_translations_post: {
            /** Model */
            model: string;
            /** Prompt */
            prompt?: string | null;
            /** @default json */
            response_format: components["schemas"]["ResponseFormat"];
            /**
             * Temperature
             * @default 0
             */
            temperature: number;
            /**
             * File
             * Format: binary
             */
            file: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}